+++ sink_writetofile_test.go
@@ -198,7 +198,10 @@ func TestWriteToFileWritesToTheTraceOutputWhenInList(t *testing.T) {
 	defer ExecPipeline(RmFile(tmpFilename))
 
 	expectedResult := `+ CatFile("./testdata/truncatefile/content.txt")
-=> CatFile("./testdata/truncatefile/content.txt")
++ => CatFile("./testdata/truncatefile/content.txt")
++ p.Stdout> This is a file of test data.
++ p.Stdout> ` + "" + `
++ p.Stdout> We copy the contents of this file to other files, as part of our testing.
 + WriteToFile("$1")
 + => WriteToFile("` + tmpFilename + `")
 + file> This is a file of test data.
@@ -240,7 +243,10 @@ func TestWriteToFileWritesToTheTraceOutputWhenInPipeline(t *testing.T) {
 	defer ExecPipeline(RmFile(tmpFilename))
 
 	expectedResult := `+ CatFile("./testdata/truncatefile/content.txt")
-=> CatFile("./testdata/truncatefile/content.txt")
++ => CatFile("./testdata/truncatefile/content.txt")
++ p.Stdout> This is a file of test data.
++ p.Stdout> ` + "" + `
++ p.Stdout> We copy the contents of this file to other files, as part of our testing.
 + WriteToFile("$1")
 + => WriteToFile("` + tmpFilename + `")
 + file> This is a file of test data.