+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - `AppendToFile()` now supports tracing
   - `AppendToTempFile()` now supports tracing
   - `Basename()` now supports tracing
+  - `CatFile()` now supports tracing
   - `Chmod()` now supports tracing
   - `CountLines()` now supports tracing
   - `CountWords()` now supports tracing